Extract server app factory and add API tests
